feat(07-05): move description block helper to shared lib
- Move pure descriptionBlocks helper out of operations context
- Point MarkdownDocument at shared/lib helper
- Preserve paragraph splitting behavior with shared/lib tests
