+++ apps/web/src/shared/lib/job-description-blocks.ts
